git-svn-id: https://ipvs.informatik.uni-stuttgart.de/SGpp/repos/branches/SGppAutoTune@4283 4eea3252-f0fb-4393-894d-40516dce545b

+++ base/src/sgpp/base/grid/generation/SquareRootGridGenerator.hpp
@@ -32,9 +32,9 @@ namespace SGPP {
          */
         virtual ~SquareRootGridGenerator();
 
-        virtual void regular(int level);
-        virtual void cliques(int level, size_t clique_size);
-        virtual void full(int level) {};
+        virtual void regular(size_t level);
+        virtual void cliques(size_t level, size_t clique_size);
+        virtual void full(size_t level) {};
         virtual void refine(RefinementFunctor* func) {};
         virtual size_t getNumberOfRefinablePoints() {
           return 0;
@@ -46,8 +46,8 @@ namespace SGPP {
           return 0;
         };
 
-        virtual void refineMaxLevel(RefinementFunctor* func, int maxLevel) {};
-        virtual size_t getNumberOfRefinablePointsToMaxLevel(int maxLevel) {
+        virtual void refineMaxLevel(RefinementFunctor* func, size_t maxLevel) {};
+        virtual size_t getNumberOfRefinablePointsToMaxLevel(size_t maxLevel) {
           return 0;
         };
